# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    The Quack dashboard API.

    Every error status (400 and up) carries the error envelope, whatever the
    handler wrote. Unknown paths and wrong methods are 404s.

    The dashboard authenticates with the HttpOnly session cookie set by Discord
    sign-in (quack_session by default). Writes made with the cookie must come from
    an allowed Origin and echo the CSRF cookie (quack_csrf by default, also
    returned as csrf_token by /auth/me) in X-CSRF-Token. A bearer session token is
    accepted instead of the cookie and skips the CSRF check.

    Writes marked with an Idempotency-Key header are safe to retry with the same
    key: a completed request is replayed with Idempotency-Replayed: true, one
    still running is a 409 with Retry-After, and reusing a key for a different
    request is a 409.

    This file is generated from the Go route table. Do not edit it; run
    "go generate ./..." from apps/backend.
  title: Quack HTTP API
  version: "5"
paths:
...
  /guilds/{discordGuildID}/modules/general-logging/repair-channel/{channelID}:
    post:
      operationId: repairLoggingChannel
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      - in: path
        name: channelID
        required: true
        schema:
          type: string
      - description: Makes the write safe to retry; reuse the key only for the identical
          request.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          description: Makes the write safe to retry; reuse the key only for the identical
            request.
          maxLength: 256
          type: string
      - description: The CSRF cookie's value. Required with the session cookie; not
          with a bearer token.
        in: header
        name: X-CSRF-Token
        schema:
          description: The CSRF cookie's value. Required with the session cookie;
            not with a bearer token.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggingSettingsPayload'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Conflict
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: Remove every route to a deleted channel; needs Manage Guild
      tags:
      - general-logging
  /guilds/{discordGuildID}/modules/general-logging/settings:
    get:
      operationId: getLoggingSettings
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggingSettingsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: General logging settings and health
      tags:
      - general-logging
    put:
      operationId: updateLoggingSettings
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      - description: Makes the write safe to retry; reuse the key only for the identical
          request.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          description: Makes the write safe to retry; reuse the key only for the identical
            request.
          maxLength: 256
          type: string
      - description: The CSRF cookie's value. Required with the session cookie; not
          with a bearer token.
        in: header
        name: X-CSRF-Token
        schema:
          description: The CSRF cookie's value. Required with the session cookie;
            not with a bearer token.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoggingSettingsPayload'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggingSettingsPayload'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Conflict
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: Replace general logging settings; needs Manage Guild
      tags:
      - general-logging
  /guilds/{discordGuildID}/modules/general-logging/status:
    get:
      operationId: getLoggingStatus
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggingStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: General logging health
      tags:
      - general-logging
...
  /guilds/{discordGuildID}/settings:
    get:
      operationId: getSettings
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSettingsEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: The guild's settings
      tags:
      - settings
    patch:
      operationId: updateSettings
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      - description: Makes the write safe to retry; reuse the key only for the identical
          request.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          description: Makes the write safe to retry; reuse the key only for the identical
            request.
          maxLength: 256
          type: string
      - description: The CSRF cookie's value. Required with the session cookie; not
          with a bearer token.
        in: header
        name: X-CSRF-Token
        schema:
          description: The CSRF cookie's value. Required with the session cookie;
            not with a bearer token.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuackGuildSettingsInput'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSettingsEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Conflict
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: Change some of the guild's settings
      tags:
      - settings
...
components:
  schemas:
...
    ApiErrorCode:
      enum:
      - validation_failed
      - authentication_required
      - reauthentication_required
      - authorization_denied
      - not_found
      - conflict
      - rate_limited
      - csrf_rejected
      - origin_rejected
      - body_too_large
      - dependency_unavailable
      - internal_error
      type: string
    ApiErrorDetail:
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        correlation_id:
          type: string
        message:
          type: string
        request_id:
          type: string
      required:
      - code
      - message
      - request_id
      - correlation_id
      type: object
    ApiErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorDetail'
      required:
      - error
      type: object
...
    ApiSettingsEnvelope:
      properties:
        settings:
          $ref: '#/components/schemas/QuackGuildSettingsResponse'
      required:
      - settings
      type: object
...
    LoggingSettings:
      properties:
        cache_entries_per_guild:
          type: integer
        channels:
          additionalProperties:
            type: string
          type:
          - object
          - "null"
        include_attachment_metadata:
          type: boolean
        include_embed_metadata:
          type: boolean
        include_message_content:
          type: boolean
        max_delivery_attempts:
          type: integer
      required:
      - channels
      - include_message_content
      - include_attachment_metadata
      - include_embed_metadata
      - cache_entries_per_guild
      - max_delivery_attempts
      type: object
    LoggingSettingsPayload:
      properties:
        enabled:
          type: boolean
        settings:
          $ref: '#/components/schemas/LoggingSettings'
      required:
      - enabled
      - settings
      type: object
    LoggingSettingsResponse:
      properties:
        enabled:
          type: boolean
        settings:
          $ref: '#/components/schemas/LoggingSettings'
        status:
          $ref: '#/components/schemas/LoggingStatus'
      required:
      - enabled
      - settings
      - status
      type: object
    LoggingStatus:
      properties:
        cached_messages:
          type: integer
        delivered:
          minimum: 0
          type: integer
        failed:
          minimum: 0
          type: integer
        last_error:
          type: string
        last_failure_at:
          format: date-time
          type:
          - "null"
          - string
      required:
      - delivered
      - failed
      - cached_messages
      type: object
    LoggingStatusResponse:
      properties:
        enabled:
          type: boolean
        status:
          $ref: '#/components/schemas/LoggingStatus'
      required:
      - enabled
      - status
      type: object
...
    QuackGuildSettingsInput:
      properties:
        appeal_queue_channel_discord_id:
          type:
          - "null"
          - string
        appeal_rejoin_url:
          type:
          - "null"
          - string
        appeal_review_reason_required:
          type:
          - "null"
          - boolean
        audit_mirror_channel_discord_id:
          type:
          - "null"
          - string
        general_logging_enabled:
          type:
          - "null"
          - boolean
        honeypot_enabled:
          type:
          - "null"
          - boolean
        managed_evidence_channel_discord_id:
          type:
          - "null"
          - string
        notification_footer:
          type:
          - "null"
          - string
        notification_introduction:
          type:
          - "null"
          - string
        tickets_enabled:
          type:
          - "null"
          - boolean
      type: object
    QuackGuildSettingsResponse:
      properties:
        appeal_queue_channel_discord_id:
          type: string
        appeal_rejoin_url:
          type: string
        appeal_review_reason_required:
          type: boolean
        audit_mirror_channel_discord_id:
          type: string
        general_logging_enabled:
          type: boolean
        guild_id:
          type: string
        honeypot_enabled:
          type: boolean
        id:
          type: string
        managed_evidence_channel_discord_id:
          type: string
        notification_footer:
          type: string
        notification_introduction:
          type: string
        starter_policy_notice_acknowledged_at:
          format: date-time
          type:
          - "null"
          - string
        starter_policy_review_required:
          type: boolean
        starter_policy_template_id:
          type: string
        tickets_enabled:
          type: boolean
      type: object
...
  securitySchemes:
    metricsKey:
      description: The metrics key (api.metrics_token).
      in: header
      name: X-Quack-Metrics-Key
      type: apiKey
    opsKey:
      description: The operator key (api.ops_token).
      in: header
      name: X-Quack-Ops-Key
      type: apiKey
    sessionBearer:
      description: The session ID as a bearer token, for callers that are not browsers.
      scheme: bearer
      type: http
    sessionCookie:
      description: The session cookie set by Discord sign-in. Its name is configurable
        (auth.session_cookie_name).
      in: cookie
      name: quack_session
      type: apiKey